+++ src/app/src/main/kotlin/org/vaccineimpact/orderlyweb/EndpointDefinition.kt
@@ -1,6 +1,6 @@
 package org.vaccineimpact.orderlyweb
 
-import org.vaccineimpact.orderlyweb.security.PermissionRequirement
+import org.vaccineimpact.orderlyweb.models.PermissionRequirement
 import spark.route.HttpMethod
 import kotlin.reflect.KClass
