+++ src/app/src/main/kotlin/org/vaccineimpact/orderlyweb/db/OnetimeTokenStore.kt
@@ -1,4 +1,4 @@
-package org.vaccineimpact.orderlyweb.security
+package org.vaccineimpact.orderlyweb.db
 
 interface OnetimeTokenStore
 {
