+++ src/app/src/main/kotlin/org/vaccineimpact/orderlyweb/db/TokenStore.kt
@@ -1,7 +1,5 @@
 package org.vaccineimpact.orderlyweb.db
 
-import org.vaccineimpact.orderlyweb.security.OnetimeTokenStore
-
 object TokenStore
 {
     val instance: OnetimeTokenStore = InMemoryTokenStore() //SQLiteTokenStore()
