--- src/app/src/main/kotlin/org/vaccineimpact/orderlyweb/security/OrderlyWebToken.kt
@@ -1,3 +1,0 @@
-package org.vaccineimpact.orderlyweb.security
-
-data class OrderlyWebToken(val username: String)
