+++ src/app/src/main/kotlin/org/vaccineimpact/orderlyweb/security/Profile.kt
@@ -4,30 +4,14 @@ import org.pac4j.core.profile.CommonProfile
 import org.vaccineimpact.orderlyweb.models.permissions.PermissionSet
 import org.vaccineimpact.orderlyweb.models.permissions.ReifiedPermission
 
-private const val USER_OBJECT = "userObject"
 private const val MISSING_PERMISSIONS = "missingPermissions"
-private const val PERMISSIONS = "montaguPermissions"
+private const val PERMISSIONS = "orderlyWebPermissions"
 private const val MISMATCHED_URL = "mismatchedURL"
-
-var CommonProfile.internalUser: InternalUser?
-    get()
-    {
-        val user = this.getAttribute(USER_OBJECT)
-        return if (user != null && user is InternalUser)
-        {
-            user
-        }
-        else
-        {
-            null
-        }
-    }
-    set(value) = this.addAttribute(USER_OBJECT, value)
 
 val CommonProfile.missingPermissions: MutableSet<ReifiedPermission>
     get() = this.getAttributeOrDefault(MISSING_PERMISSIONS, default = mutableSetOf())
 
-var CommonProfile.montaguPermissions: PermissionSet
+var CommonProfile.orderlyWebPermissions: PermissionSet
     get() = this.getAttributeOrDefault(PERMISSIONS, PermissionSet())
     set(value) = this.addAttribute(PERMISSIONS, value)
