--- src/app/src/main/kotlin/org/vaccineimpact/orderlyweb/security/TokenVerifier.kt
@@ -1,8 +1,0 @@
-package org.vaccineimpact.orderlyweb.security
-
-import java.security.interfaces.RSAPublicKey
-
-class TokenVerifier(publicKey: RSAPublicKey, val expectedIssuer: String)
-{
-    val signatureConfiguration = PublicKeyOnlySignatureConfiguration(publicKey)
-}
