+++ src/app/src/main/kotlin/org/vaccineimpact/orderlyweb/security/WebTokenHelper.kt
@@ -1,14 +1,16 @@
 package org.vaccineimpact.orderlyweb.security
 
 import org.vaccineimpact.orderlyweb.db.AppConfig
+import org.vaccineimpact.orderlyweb.security.authentication.TokenVerifier
+import org.vaccineimpact.orderlyweb.security.issuing.KeyHelper
+import org.vaccineimpact.orderlyweb.security.issuing.TokenIssuer
 import java.security.KeyPair
-import java.security.interfaces.RSAPublicKey
 
 class WebTokenHelper(keyPair: KeyPair,
                      val issuerName: String)
 {
     val issuer = TokenIssuer(keyPair, issuerName)
-    val verifier = TokenVerifier(keyPair.public as RSAPublicKey, issuerName)
+    val verifier = TokenVerifier(keyPair, issuerName)
 
     companion object
     {
