+++ src/app/src/main/kotlin/org/vaccineimpact/orderlyweb/security/authentication/TokenVerifier.kt
@@ -1,0 +1,9 @@
+package org.vaccineimpact.orderlyweb.security.authentication
+
+import org.pac4j.jwt.config.signature.RSASignatureConfiguration
+import java.security.KeyPair
+
+class TokenVerifier(keyPair: KeyPair, val expectedIssuer: String)
+{
+    val signatureConfiguration = RSASignatureConfiguration(keyPair)
+}
