+++ src/app/src/main/kotlin/org/vaccineimpact/orderlyweb/security/authorization/OrderlyWebAuthorizer.kt
@@ -1,4 +1,4 @@
-package org.vaccineimpact.orderlyweb.security
+package org.vaccineimpact.orderlyweb.security.authorization
 
 import org.pac4j.core.authorization.authorizer.AbstractRequireAllAuthorizer
 import org.pac4j.core.context.WebContext
@@ -6,6 +6,10 @@ import org.pac4j.core.profile.CommonProfile
 import org.pac4j.sparkjava.SparkWebContext
 import org.slf4j.LoggerFactory
 import org.vaccineimpact.orderlyweb.DirectActionContext
+import org.vaccineimpact.orderlyweb.models.PermissionRequirement
+import org.vaccineimpact.orderlyweb.security.mismatchedURL
+import org.vaccineimpact.orderlyweb.security.missingPermissions
+import org.vaccineimpact.orderlyweb.security.orderlyWebPermissions
 
 open class OrderlyWebAuthorizer(requiredPermissions: Set<PermissionRequirement>)
     : AbstractRequireAllAuthorizer<PermissionRequirement, CommonProfile>()
@@ -47,7 +51,7 @@ open class OrderlyWebAuthorizer(requiredPermissions: Set<PermissionRequirement>)
 
     override fun check(context: WebContext, profile: CommonProfile, element: PermissionRequirement): Boolean
     {
-        val profilePermissions = profile.montaguPermissions
+        val profilePermissions = profile.orderlyWebPermissions
         val reifiedRequirement = element.reify(DirectActionContext(context as SparkWebContext))
 
         val hasPermission = profilePermissions.any { reifiedRequirement.satisfiedBy(it) }
