+++ src/app/src/main/kotlin/org/vaccineimpact/orderlyweb/security/clients/GithubDirectClient.kt
@@ -1,4 +1,4 @@
-package org.vaccineimpact.orderlyweb.security
+package org.vaccineimpact.orderlyweb.security.clients
 
 import org.eclipse.egit.github.core.client.GitHubClient
 import org.pac4j.core.client.DirectClient
@@ -10,6 +10,8 @@ import org.pac4j.core.profile.CommonProfile
 import org.vaccineimpact.orderlyweb.db.OrderlyUserRepository
 import org.vaccineimpact.orderlyweb.models.ErrorInfo
 import org.vaccineimpact.orderlyweb.models.permissions.PermissionSet
+import org.vaccineimpact.orderlyweb.security.authentication.GithubAuthenticator
+import org.vaccineimpact.orderlyweb.security.orderlyWebPermissions
 
 class GithubDirectClient : DirectClient<TokenCredentials, CommonProfile>(), OrderlyWebTokenCredentialClient
 {
@@ -29,7 +31,7 @@ class GithubDirectClient : DirectClient<TokenCredentials, CommonProfile>(), Orde
 
     private fun addLoginPermission(profile: CommonProfile): CommonProfile
     {
-        profile.montaguPermissions = PermissionSet("*/can-login")
+        profile.orderlyWebPermissions = PermissionSet("*/can-login")
         return profile
     }
 }
