+++ src/app/src/main/kotlin/org/vaccineimpact/orderlyweb/security/clients/JWTCookieClient.kt
@@ -1,8 +1,11 @@
-package org.vaccineimpact.orderlyweb.security
+package org.vaccineimpact.orderlyweb.security.clients
 
 import org.pac4j.core.context.WebContext
 import org.pac4j.http.client.direct.CookieClient
 import org.vaccineimpact.orderlyweb.models.ErrorInfo
+import org.vaccineimpact.orderlyweb.security.authentication.OrderlyWebBearerTokenAuthenticator
+import org.vaccineimpact.orderlyweb.security.authentication.TokenVerifier
+import org.vaccineimpact.orderlyweb.security.extractPermissionsFromToken
 
 // This client receives the token as TokenCredentials and stores the result as JwtProfile
 class JWTCookieClient(helper: TokenVerifier): OrderlyWebTokenCredentialClient, CookieClient(
@@ -23,6 +26,6 @@ class JWTCookieClient(helper: TokenVerifier): OrderlyWebTokenCredentialClient, C
 
     override val errorInfo = ErrorInfo(
             "cookie-bearer-token-invalid",
-            "Bearer token not supplied in cookie '${JWTCookieClient.cookie}', or bearer token was invalid"
+            "Bearer token not supplied in cookie '$cookie', or bearer token was invalid"
     )
 }
