+++ src/app/src/main/kotlin/org/vaccineimpact/orderlyweb/security/clients/JWTParameterClient.kt
@@ -1,8 +1,12 @@
-package org.vaccineimpact.orderlyweb.security
+package org.vaccineimpact.orderlyweb.security.clients
 
 import org.pac4j.core.context.WebContext
 import org.pac4j.http.client.direct.ParameterClient
+import org.vaccineimpact.orderlyweb.db.OnetimeTokenStore
 import org.vaccineimpact.orderlyweb.models.ErrorInfo
+import org.vaccineimpact.orderlyweb.security.*
+import org.vaccineimpact.orderlyweb.security.authentication.OrderlyWebOnetimeTokenAuthenticator
+import org.vaccineimpact.orderlyweb.security.authentication.TokenVerifier
 
 // This client receives the token as TokenCredentials and stores the result as JwtProfile
 class JWTParameterClient(helper: TokenVerifier, tokenStore: OnetimeTokenStore) : OrderlyWebTokenCredentialClient, ParameterClient(
